+++ src/main/kotlin/com/example/springoraclehibernate/services/PhoneTypeService.kt
@@ -1,11 +1,12 @@
 package com.example.springoraclehibernate.services
 
 import com.example.springoraclehibernate.domain.PhoneType
+import java.math.BigInteger
 
 interface PhoneTypeService {
     fun getAllTypes() : List<PhoneType>
-    fun getTypeById(id: Long) : PhoneType?
+    fun getTypeById(id: BigInteger) : PhoneType?
     fun addType(type: PhoneType) : PhoneType
-    fun deleteType(id: Long)
+    fun deleteType(id: BigInteger)
     fun updateType(type: PhoneType) : PhoneType
 }
